Remove old services; move store content to relevant folders
