+++ src/app/game/alliance/alliance.actions.ts
