+++ src/app/game/alliance/alliance.reducer.ts
