+++ src/app/game/map/map.actions.ts
