+++ src/app/game/map/map.reducer.ts
