+++ src/app/game/player/player.reducer.ts
