+++ src/app/game/report/report-item/report-item.component.html
