+++ src/app/game/report/report.actions.ts
