+++ src/app/game/report/report.effects.ts
