+++ src/app/game/services/index.ts
@@ -1,5 +1,3 @@
 export { SocketService } from './socket.service';
-export { TownService } from './town.service';
 export { MapService } from './map.service';
 export { CommandService } from './command.service';
-export { ReportService } from './report.service';
