--- src/app/game/services/town.service.spec.ts
@@ -1,16 +1,0 @@
-// /* tslint:disable:no-unused-variable */
-
-// import { addProviders, async, inject } from '@angular/core/testing';
-// import { RestaurantService } from './restaurant.service';
-
-// describe('Service: Restaurant', () => {
-//   beforeEach(() => {
-//     addProviders([RestaurantService]);
-//   });
-
-//   it('should ...',
-//     inject([RestaurantService],
-//       (service: RestaurantService) => {
-//         expect(service).toBeTruthy();
-//       }));
-// });
