+++ src/app/game/town/town.model.ts
