+++ src/app/game/town/town.reducer.ts
